Added a website coming soon splash page
A temporary splash page has been added to this project to foretell of a coming website. As for now, a home page, CSS, and JS file has been added to support this splash page. The System.Web.Optimization package has been installed, along with several supporting packages. This allows the use of @Script and @Style in the Razar markup.